remove the Additional TestHelper and modify the AttachmentServiceImpl
