+++ api/src/test/java/org/openmrs/module/attachments/obs/TestHelper.java
@@ -4,11 +4,18 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
+import java.util.Locale;
 import java.util.Random;
 
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.RandomStringUtils;
+import org.openmrs.ConceptClass;
+import org.openmrs.ConceptComplex;
+import org.openmrs.ConceptDatatype;
+import org.openmrs.ConceptDescription;
+import org.openmrs.ConceptName;
 import org.openmrs.Encounter;
 import org.openmrs.EncounterType;
 import org.openmrs.GlobalProperty;
@@ -22,6 +29,7 @@ import org.openmrs.module.attachments.AttachmentsActivator;
 import org.openmrs.module.attachments.AttachmentsConstants;
 import org.openmrs.module.attachments.AttachmentsContext;
 import org.openmrs.module.attachments.ComplexObsSaver;
+import org.openmrs.obs.handler.BinaryDataHandler;
 import org.openmrs.ui.framework.WebConstants;
 import org.openmrs.util.OpenmrsConstants;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -61,6 +69,8 @@ public class TestHelper {
 	        "application/octet-stream", "mock_content".getBytes());
 	
 	protected MockMultipartFile lastSavedMultipartImageFile;
+	
+	protected ConceptComplex ConceptComplexOutAttach;
 	
 	/**
 	 * @return The last saved test image file, null if none was ever saved.
@@ -117,6 +127,24 @@ public class TestHelper {
 		
 		context.getAdministrationService().saveGlobalProperty(
 		    new GlobalProperty(AttachmentsConstants.GP_ENCOUNTER_TYPE_UUID, AttachmentsConstants.ENCOUNTER_TYPE_UUID));
+		
+		if (context.getConceptService().getConceptByName("OutOfAttachmentsTestComplex") == null) {
+			ConceptComplex conceptComplexOutAttach = new ConceptComplex();
+			conceptComplexOutAttach.setHandler(BinaryDataHandler.class.getSimpleName());
+			ConceptName conceptName = new ConceptName("OutOfAttachmentsTestComplex", Locale.ENGLISH);
+			conceptComplexOutAttach.setFullySpecifiedName(conceptName);
+			conceptComplexOutAttach.setPreferredName(conceptName);
+			conceptComplexOutAttach
+			        .setConceptClass(context.getConceptService().getConceptClassByUuid(ConceptClass.QUESTION_UUID));
+			conceptComplexOutAttach
+			        .setDatatype(context.getConceptService().getConceptDatatypeByUuid(ConceptDatatype.COMPLEX_UUID));
+			conceptComplexOutAttach
+			        .addDescription(new ConceptDescription("Out-of-Attachments test concept complex", Locale.ENGLISH));
+			context.getConceptService().saveConcept(conceptComplexOutAttach);
+			ConceptComplexOutAttach = conceptComplexOutAttach;
+			// System.out.println(conceptComplexOutAttach.getUuid());
+			
+		}
 	}
 	
 	public Obs getTestComplexObs() throws IOException {
@@ -232,6 +260,150 @@ public class TestHelper {
 			obsList.add(obsSaver.saveOtherAttachment(visit, patient, encounter, fileCaption, multipartRandomFile,
 			    ValueComplex.INSTRUCTIONS_DEFAULT));
 		}
+		return obsList;
+	}
+	
+	/**
+	 * Boilerplate method to save a collection of complex obs.
+	 *
+	 * @param count The number of the complex obs to be saved.
+	 */
+	public List<Obs> saveComplexObsForEncounter(int count) throws IOException {
+		init();
+		
+		// Creating the list of Complex Obs relevnt to the attachment module (
+		// Attachments ) using ComplexObsSaver class.
+		List<Obs> obsList = new ArrayList<>();
+		byte[] randomData = new byte[20];
+		
+		Patient patient = context.getPatientService().getPatient(2);
+		Visit visit = context.getVisitService().getVisit(1);
+		
+		EncounterService encounterService = context.getEncounterService();
+		EncounterType encounterType = encounterService.getEncounterType(1);
+		Provider provider = context.getProviderService().getProvider(1);
+		context.getAdministrationService().saveGlobalProperty(
+		    new GlobalProperty(AttachmentsConstants.GP_ENCOUNTER_TYPE_UUID, encounterType.getUuid()));
+		Encounter encounter = context.getAttachmentEncounter(patient, visit, provider);
+		
+		for (int i = 0; i < count; i++) {
+			String fileCaption = RandomStringUtils.randomAlphabetic(12);
+			new Random().nextBytes(randomData);
+			MockMultipartFile multipartRandomFile = new MockMultipartFile(String.valueOf(i), String.valueOf(i),
+			        "application/octet-stream", randomData);
+			obsList.add(obsSaver.saveOtherAttachment(visit, patient, encounter, fileCaption, multipartRandomFile,
+			    ValueComplex.INSTRUCTIONS_DEFAULT));
+		}
+		{
+			// Creating the complex Obs not relevent / relevent to the attachment module
+			// during the same encounter.
+			Obs obs = new Obs();
+			// set out-of-Attachment test concept complex
+			obs.setConcept(ConceptComplexOutAttach);
+			
+			// set Attachement concept complex
+			// ConceptComplex conceptComplex = context.getConceptService()
+			// .getConceptComplex(obsList.get(0).getConcept().getConceptId());
+			// obs.setConcept(conceptComplex);
+			
+			obs.setObsDatetime(new Date());
+			obs.setPerson(patient);
+			obs.setEncounter(encounter);
+			
+			new Random().nextBytes(randomData);
+			MockMultipartFile multipartRandomFile = new MockMultipartFile("1", "1", "application/octet-stream", randomData);
+			obs.setComplexData(
+			    complexDataHelper.build(ValueComplex.INSTRUCTIONS_DEFAULT, multipartRandomFile.getOriginalFilename(),
+			        multipartRandomFile.getBytes(), multipartRandomFile.getContentType()).asComplexData());
+			obs = context.getObsService().saveObs(obs, null);
+			
+			// Add complex obs to ObsList when Obs is created using attachment concept
+			// complex
+			// obsList.add(obs);
+		}
+		{
+			// Createing some other obs ( not complex Obs ) during the same encounter
+			Obs otherObs = new Obs();
+			otherObs.setConcept(context.getConceptService().getConcept(3));
+			otherObs.setObsDatetime(new Date());
+			otherObs.setEncounter(encounter);
+			otherObs.setPerson(patient);
+			otherObs.setValueText("Some text value for a test obs.");
+			otherObs = context.getObsService().saveObs(otherObs, null);
+		}
+		// Only return the attachment complex obs related to the attachment module.
+		return obsList;
+	}
+	
+	public List<Obs> saveComplexObsForVisit(int count) throws IOException {
+		init();
+		
+		List<Obs> obsList = new ArrayList<>();
+		byte[] randomData = new byte[20];
+		
+		Patient patient = context.getPatientService().getPatient(2);
+		Visit visit = context.getVisitService().getVisit(1);
+		EncounterService encounterService = context.getEncounterService();
+		EncounterType encounterType = encounterService.getEncounterType(1);
+		Provider provider = context.getProviderService().getProvider(1);
+		
+		context.getAdministrationService().saveGlobalProperty(
+		    new GlobalProperty(AttachmentsConstants.GP_ENCOUNTER_TYPE_UUID, encounterType.getUuid()));
+		
+		Encounter encounter = context.getAttachmentEncounter(patient, visit, provider);
+		Encounter encounter2 = context.getAttachmentEncounter(patient, visit, provider);
+		
+		// saving first obs different from the others
+		String fileCaption = RandomStringUtils.randomAlphabetic(12);
+		new Random().nextBytes(randomData);
+		MockMultipartFile multipartRandomFile = new MockMultipartFile(String.valueOf(0), String.valueOf(0),
+		        "application/octet-stream", randomData);
+		obsList.add(obsSaver.saveOtherAttachment(visit, patient, encounter, fileCaption, multipartRandomFile,
+		    ValueComplex.INSTRUCTIONS_DEFAULT));
+		
+		for (int i = 1; i < count; i++) {
+			fileCaption = RandomStringUtils.randomAlphabetic(12);
+			new Random().nextBytes(randomData);
+			multipartRandomFile = new MockMultipartFile(String.valueOf(i), String.valueOf(i), "application/octet-stream",
+			        randomData);
+			obsList.add(obsSaver.saveOtherAttachment(visit, patient, encounter2, fileCaption, multipartRandomFile,
+			    ValueComplex.INSTRUCTIONS_DEFAULT));
+		}
+		{
+			// Creating the complex Obs not relevent / relevent to the attachment module
+			// during the same visit.
+			Obs obs = new Obs();
+			// set out-of-Attachment test concept complex
+			obs.setConcept(ConceptComplexOutAttach);
+			
+			// // set Attachement concept complex
+			// ConceptComplex conceptComplex = context.getConceptService()
+			// .getConceptComplex(obsList.get(0).getConcept().getConceptId());
+			// obs.setConcept(conceptComplex);
+			
+			obs.setObsDatetime(new Date());
+			obs.setPerson(patient);
+			obs.setEncounter(encounter);
+			obs.setComplexData(
+			    complexDataHelper.build(ValueComplex.INSTRUCTIONS_DEFAULT, multipartRandomFile.getOriginalFilename(),
+			        multipartRandomFile.getBytes(), multipartRandomFile.getContentType()).asComplexData());
+			obs = context.getObsService().saveObs(obs, null);
+			
+			// Add complex obs to ObsList when Obs is created using attachment concept
+			// complex
+			// obsList.add(obs);
+		}
+		{
+			// Create some other Obs ( not complex Obs ) during the same visit .
+			Obs otherObs = new Obs();
+			otherObs.setConcept(context.getConceptService().getConcept(3));
+			otherObs.setObsDatetime(new Date());
+			otherObs.setEncounter(encounter2);
+			otherObs.setPerson(patient);
+			otherObs.setValueText("Some text value for a test obs.");
+			otherObs = context.getObsService().saveObs(otherObs, null);
+		}
+		
 		return obsList;
 	}
 }
